+++ CalorieCalculator/CalculateFatCalories.cs
@@ -15,7 +15,7 @@
         /// </summary>
         /// <param name="fatIntake">資質の摂取量</param>
         /// <returns>資質の摂取量から計算したカロリー</returns>
-        public long Execute(double fatIntake)
+        public long Execute(decimal fatIntake)
         {
             var fat = Math.Round(fatIntake, 1, MidpointRounding.AwayFromZero);
