+++ CalorieCalculatorTests/CalorieCalculationTests.cs
@@ -12,7 +12,7 @@ namespace CalorieCalculatorTests
         /// </summary>
         [Theory]
         [MemberData(nameof(CorrectCaloriesData))]
-        public void CorrectCalories(double protein, double fat, double carbohydrate, double expected)
+        public void CorrectCalories(decimal protein, decimal fat, decimal carbohydrate, decimal expected)
         {
             // Arrange
             var amountOfIntake = new AmountOfIntake
@@ -28,8 +28,8 @@ namespace CalorieCalculatorTests
 
             // Act
             var result = calculateProteinCalories.Execute(amountOfIntake.Protein) +
-                            caluculateFatCalories.Execute(amountOfIntake.Fat) +
-                            calculateCarbohydrateCalories.Execute(amountOfIntake.Carbohydrate);
+                                caluculateFatCalories.Execute(amountOfIntake.Fat) +
+                                calculateCarbohydrateCalories.Execute(amountOfIntake.Carbohydrate);
 
             // Assert
             Assert.Equal(expected, result);
@@ -38,27 +38,26 @@ namespace CalorieCalculatorTests
         /// <summary>
         /// CorrectCaloriesのテストデータ
         /// </summary>
-        public static TheoryData<double, double, double, long> CorrectCaloriesData =>
-            new TheoryData<double, double, double, long>
+        public static TheoryData<decimal, decimal, decimal, long> CorrectCaloriesData =>
+            new TheoryData<decimal, decimal, decimal, long>
             {
                 // 計算が正しく行えているか
-                { 2.5, 0.3, 37.1, 161 },
-                { 16.5, 10.0, 12.1, 204 },
+                { 2.5M, 0.3M, 37.1M, 161 },
+                { 16.5M, 10.0M, 12.1M, 204 },
                 // Proteinの四捨五入テスト
-                { 0.64, 0.0, 0.0, 2 },
-                { 0.65, 0.0, 0.0, 3 },
+                { 0.64M, 0.0M, 0.0M, 2 },
+                { 0.65M, 0.0M, 0.0M, 3 },
                 // Fatの四捨五入テスト+結果の四捨五入テスト
-                { 0.0, 0.54, 0.0, 5 },
-                { 0.0, 0.55, 0.0, 5 },
+                { 0.0M, 0.54M, 0.0M, 5 },
+                { 0.0M, 0.55M, 0.0M, 5 },
                 // Carbonydrateの四捨五入テスト
-                { 0.0, 0.0, 0.64, 2 },
-                { 0.0, 0.0, 0.65, 3 },
+                { 0.0M, 0.0M, 0.64M, 2 },
+                { 0.0M, 0.0M, 0.65M, 3 },
                 // TODO: 最大値と最小値の要件が不明なので判明したらテストを追加する
                 // 最小値のテスト（0以下はない前提）
-                { 0.0, 0.0, 0.0, 0 },
+                { 0.0M, 0.0M, 0.0M, 0 },
                 // 最大値のテスト（仮に7桁を最大とする）
-                { 9999999.9, 9999999.9, 9999999.9, 169999999 }
-
+                { 9999999.9M, 9999999.9M, 9999999.9M, 169999999 }
             };
     }
 }
